Finished Modal Layout
I finished the general layout of the website. All that's left is to populate it with content and a final styling of aesthetics.

Also finally uploaded the JSON, forgot to upload the public folder So prior versions to this should not work unless you take this public folder and replace it with the prior versions' public folder.