New Component: Img/Video
The new component allows for videos to be part of the slideshow. Using the project image as a thumbnail, the component will switch out the element for a the appropiate media element. Currently satisfied, I will probably be adjusting this component down the line. Also might have to restructure the JSON file to include a property for thumbnail.