Reworked Slideshow Style
Switched most the styling away from the Grid display type. It would be a pain to create the responsive stylings with the amount of grids I used.  This process had me moving around some HTML code and removing redundant classes/ids. I also moved the nav buttons to be absolutely posistioned on both ends of the slideshow. This is the main setup for reworking the nav bullets to include animation as well as animating the slideshow itself. Not fully set on the overall styling of the section and the website.