SASS conversion
Migrated the CSS to very basic SASS.

I've considered this portfolio website as a playground for my front-end development skills. Current standards include knowing a CSS preprocessor . I'm looking to make the most of it and see how well I can combine it with React.

For saftey i've left the old CSS files in this commit, they'll be removed in the next one as I tended to include more SASS only functions.