Functional slideshow animation
The slideshow now "slides", it is mostly functional but will require heavy tuning both in JS and SCSS.

The code was based on Ashwanth A.R's write up in medium. I've been working it into my structure so currently it's like a Frankenstien code.

I've also stopped using the Media component as I no longer want to put a video in the slide, especially due to this new aspect ratio. It may rework it for the modal.