Updated slideshow component
This is a work in progress update. The prior version, which was taken from the internet, had the slide that went off screen automatically move itself to the right off screen.
This update has the slide wait on the left off screen side until it's time to change slide, then it is prepped to go to the right. Ideally with small adjustments this code will allow the slide animation be done in reverse for decending order.

There is a lot of redundant code that needs adjusting.